Optimizing program classes.
