+++ src/asset/models/Cart.js
@@ -1,42 +1,11 @@
-class Cart {
+import Display from "./Display.js";
+
+class Cart extends Display {
   constructor(parent, price) {
-    this.parent = parent;
+    super(parent);
     this.totalPrice = price;
     this.products = [];
     this.toShow = [];
-
-    this.parent.addEventListener("click", this);
-  }
-
-  showProducts() {
-    this.toShow = [...new Set(this.products)];
-    this.parent.innerHTML = "";
-    this.toShow.forEach((product) => {
-      const qty = this.products.filter((p) => p.id === product.id).length;
-      this.createCard(product, qty);
-    });
-
-    this.calculateTotalPrice();
-  }
-
-  createCard(data, qty) {
-    const cardEle = document.createElement("div");
-
-    const imgEle = this.productImg(data);
-    const infoEle = this.productInfo(data);
-    const controlEle = this.productControl(data, qty);
-
-    cardEle.innerHTML = imgEle;
-    cardEle.innerHTML += infoEle;
-    cardEle.innerHTML += controlEle;
-
-    this.parent.appendChild(cardEle);
-  }
-
-  productImg(data) {
-    const { image, alt } = data;
-    const imgJSX = `<img src="${image}" alt="${alt}"/>`;
-    return imgJSX;
   }
 
   productInfo(data) {
